Move Step out of boot_animation.py
